+++ test/workspace/git_test.rb
@@ -105,4 +105,22 @@ class WorkspaceGitTest < Minitest::Test
         ], info)
     end
   end
+
+  def test_git_fetch_failed
+    with_workspace(**workspace_params.merge(pull_number: 999999999)) do |workspace|
+      error = assert_raises(Runners::Workspace::Git::FetchFailed) do
+        workspace.send(:prepare_head_source, nil)
+      end
+      assert_match %r{\Agit-fetch failed: fatal: couldn't find remote ref refs/pull/999999999/head}, error.message
+    end
+  end
+
+  def test_git_checkout_failed
+    with_workspace(**workspace_params.merge(head: "invalid_commit")) do |workspace|
+      error = assert_raises(Runners::Workspace::Git::CheckoutFailed) do
+        workspace.send(:prepare_head_source, nil)
+      end
+      assert_match %r{\Agit-checkout failed: error: pathspec 'invalid_commit' did not match any file\(s\) known to git}, error.message
+    end
+  end
 end